Struct eternity::multi_bucket::CachedLimitedEnums

pub struct CachedLimitedEnums<Key, Value> where
    Key: Hash + PartialEq + Clone + Eq + Send + Sync,
    Value: Clone + Send + Sync,  { /* fields omitted */ }

Disclaimer:

The Key is expected to be an enum! If your key is no enum, use LimitedRequests.

Implementation

This type uses std::mem::discriminant on the key to differentiate between bucket groups.

It’s tailored to limit calling and caching HTTP endpoints, where the the key represents the endpoints containing arguments. At first, the type will convert the key to a discriminant to pick the the bucket group, ignoring the fields, and then it will use the enum with fields to find the actual bucket instance.

Warning

As said, if the key is not an enum, this type will not work as expected.
Use LimitedRequests instead for generic keys.

Implementations

impl<Key, Value> CachedLimitedEnums<Key, Value> where
Key: Hash + PartialEq + Clone + Eq + Send + Sync,
Value: Clone + Send + Sync,

#[must_use]pub fn new() -> Self

Create an empty CachedLimitedEnums.

pub fn insert_enum(&mut self, key: Key) -> Result<(), Key> where
Key: ToBucket<Key, Value>,

Inserts a single group by turning a GroupKey to a Bucket.
Returns an Err with a GroupKey if it failed converting to None.

It requires mutable ownership over self, if you know all groups before using this data structure, insert them ahead of time and use cache_or.
This approach requires immutable access and prevents needing a locking mechanism.

If you do not know all types beforehand, you will need a locking mechanism to call mut_cache_or for the mutable ownership.
It allows you to add group buckets on the run.

Errors

If a key cannot be converted, the method will return the nonconvertible keys. This means, you have provided a key that cannot construct a bucket group.

pub fn insert_bucket(&mut self, key: &Key, bucket: Bucket<Key, Value>)

Inserts an already created bucket for a key.

pub fn build_group<'a>(
&'a mut self,
key: &'a Key
) -> CachedLimitedEnumsModifier<'a, Key, Value>

Inserts an already created bucket for a key.

Add a bucket group for the bucket_key. The method returns a new builder to add it to self.

pub fn insert_enums(
&mut self,
keys: impl IntoIterator<Item = Key>
) -> Result<(), Vec<Key>> where
Key: ToBucket<Key, Value>,

Inserts multiple enums. The type must be an enum and nothing else.

This method requires mutable ownership over self, if you know all enum before using this data structure, insert them ahead of time and use cache_or.

This approach requires immutable access and prevents needing a locking mechanism.

If you do not know all types beforehand, you will need a locking mechanism to call mut_cache_or for mutable ownership.
It allows you to add group buckets on the run.

In order to provide automatic coverage for your enum’s variants, try the strum::EnumIter macro.

Errors

If a key cannot be converted, the method will return a Vec containing nonconvertible keys. This means, you have provided a key that cannot construct a bucket group.

pub async fn hit_limit(&mut self, key: &Key) -> Option<RateLimitInfo<Value>>

This method will return a rate limit once the conditions are met. However, the method does not cache anything if the cache feature is not enabled.

Warning

The key may only be an enum and no other type.

pub async fn cache_or(
&self,
key: &Key,
function: impl Future<Output = Option<Value>> + Send
) -> Option<Value>

This method will access the cached value if rate limited or execute the given function.

Difference to mut_cache_or

When the group_key does not exist, the group won’t be created and None will be returned.

Cache Feature

If the function is executed and caching is enabled, the returned value will be cached.

Warning

The key may only be an enum and no other type.

pub async fn mut_cache_or(
&mut self,
key: &Key,
function: impl Future<Output = Option<Value>> + Send
) -> Option<Value> where
Key: ToBucket<Key, Value>,

This method will access the cached value if limited or execute the given function.

Difference to cache_or

When the group_key does not exist, the group will be created and and the function will be executed.

Cache Feature

If the function is executed and caching is enabled, the returned value will be cached.

Warning

The key may only be an enum and no other type.