Struct eternity::multi_bucket::LimitedRequests

pub struct LimitedRequests<GroupKey, BucketKey = GroupKey, Value = ()> where
    GroupKey: Hash + PartialEq + Clone + Eq + Send + Sync,
    BucketKey: Hash + PartialEq + Clone + Eq + Send + Sync,
    Value: Clone + Send + Sync,  { /* fields omitted */ }

This limiter can use a group key for each bucket group and a bucket key to get the bucket instance.

A GroupKey can be a String, while the key for each instance of a bucket can be a number.

By default, the BucketKey is the GroupKey.

Example:

Use the command name as group key and the user ID as instance key. You can use the same key for both too.

Implementations

impl<GroupKey, BucketKey, Value> LimitedRequests<GroupKey, BucketKey, Value> where
GroupKey: Hash + PartialEq + Clone + Eq + Send + Sync,
BucketKey: Hash + PartialEq + Clone + Eq + Send + Sync,
Value: Clone + Send + Sync,

#[must_use]pub fn new() -> Self

Creates an empty limiter.

pub fn insert_group(&mut self, bucket_key: GroupKey) -> Result<(), GroupKey> where
GroupKey: ToBucket<BucketKey, Value>,

Inserts a single group by turning a GroupKey to a Bucket.

It requires mutable ownership over self, if you know all groups before using this data structure, insert them ahead of time and use cache_or.
This approach requires immutable access and prevents needing a locking mechanism.

If you do not know all types beforehand, you will need a locking mechanism to call mut_cache_or for the mutable ownership.
It allows you to add group buckets on the run.

Errors

If a key cannot be converted, the method will return a Vec containing nonconvertible keys. This means, you have provided a key that cannot construct a bucket group.

pub fn build_group(
&mut self,
bucket_key: GroupKey
) -> LimitedRequestsModifier<'_, GroupKey, BucketKey, Value>

Add a bucket group for the bucket_key. The method returns a new builder to add it to self.

pub fn insert_groups<IterItem>(
&mut self,
bucket_keys: impl IntoIterator<Item = GroupKey>
) -> Result<(), Vec<GroupKey>> where
GroupKey: ToBucket<BucketKey, Value>,

Inserts multiple groups at once using GroupKeys.

This method requires mutable ownership over self, if you know all groups before using this data structure, insert them ahead of time and use cache_or.

This approach requires immutable access and prevents needing a locking mechanism.

If you do not know all types beforehand, you will need a locking mechanism to call mut_cache_or for mutable ownership.
It allows you to add group buckets on the run.

Errors

If a key cannot be converted, the method will return a Vec containing nonconvertible keys. This means, you have provided a key that cannot construct a bucket group.

pub fn insert_bucket(
&mut self,
group_key: GroupKey,
bucket: Bucket<BucketKey, Value>
)

Inserts a single bucket to the group with group_key.

This method requires mutable ownership over self, if you know all groups before using this data structure, insert them ahead of time and use cache_or.

This approach requires immutable access and prevents needing a locking mechanism.

If you do not know all types beforehand, you will need a locking mechanism to call mut_cache_or for mutable ownership.
It allows you to add group buckets on the run.

pub fn insert_buckets(
&mut self,
group_key: &GroupKey,
buckets: impl IntoIterator<Item = Bucket<BucketKey, Value>>
)

Inserts multiple buckets to the group with group_key.

This method requires mutable ownership over self, if you know all groups before using this data structure, insert them ahead of time and use cache_or.

This approach requires immutable access and prevents needing a locking mechanism.

If you do not know all types beforehand, you will need a locking mechanism to call mut_cache_or for mutable ownership.
It allows you to add group buckets on the run.

pub async fn hit_limit(
&self,
bucket_key: &GroupKey,
value_key: &BucketKey
) -> Option<RateLimitInfo<Value>>

This method will return a rate limit once the conditions are met. However, the method does not cache anything if the cache feature is not enabled.

pub async fn cache_or(
&self,
group_key: &GroupKey,
bucket_key: &BucketKey,
function: impl Future<Output = Option<Value>> + Send
) -> Option<Value>

This method will access the cached value if limited or execute the given function.

Difference to mut_cache_or

When the group_key does not exist, the group won’t be created and None will be returned.

Cache Feature

If the function is executed and caching is enabled, the returned value will be cached.

pub async fn mut_cache_or(
&mut self,
group_key: &GroupKey,
bucket_key: &BucketKey,
function: impl Future<Output = Option<Value>> + Send
) -> Option<Value> where
GroupKey: ToBucket<BucketKey, Value>,

This method will access the cached value if limited or execute the given function.

Difference to cache_or

When the group_key does not exist, the group will be created.

Cache Feature

If the function is executed and caching is enabled, the returned value will be cached.