Trait eternity::multi_bucket::ToBucket

pub trait ToBucket<Key, Value = ()> where
    Key: Hash + PartialEq + Clone + Eq + Send + Sync,
    Value: Clone + Send, {
    fn to_bucket(&self) -> Option<Bucket<Key, Value>>;
}

Creates a bucket using the key.

This trait ensures that all possible valid sets can become a bucket. The Value resembles the cached value and is optional, if caching is not needed, it can be be ignored.

Required methods

fn to_bucket(&self) -> Option<Bucket<Key, Value>>

Turns a key into a bucket.

Loading content...